# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smart Pet API",
    "version": "v1",
    "description": "Unified API for the breeder platform + owner device loop. Routes are being migrated to this spec file-by-file; anything not listed here is still served but not yet documented."
  },
  "servers": [
    {
      "url": "/api/v1"
    },
    {
      "url": "/api",
      "description": "unversioned alias (deprecated)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
      "RateLimited": {
        "description": "Too many requests. Back off per `Retry-After`.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            },
            "description": "Seconds to wait."
          },
          "RateLimit-Limit": {
            "schema": {
              "type": "integer"
            }
          },
          "RateLimit-Remaining": {
            "schema": {
              "type": "integer"
            }
          },
          "RateLimit-Reset": {
            "schema": {
              "type": "integer"
            },
            "description": "Seconds until the window resets."
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/breeder/buyers/messages/broadcast": {
      "post": {
        "tags": [
          "breeder: buyer comms"
        ],
        "summary": "Broadcast a message to a group of buyers.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "subject": {
                    "type": "string",
                    "minLength": 1
                  },
                  "body": {
                    "type": "string",
                    "minLength": 1
                  },
                  "litterId": {
                    "nullable": true,
                    "type": "string"
                  },
                  "status": {
                    "nullable": true,
                    "type": "string"
                  }
                },
                "required": [
                  "subject",
                  "body"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sent": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "sent"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/breeder/buyers/messages/{buyerId}": {
      "post": {
        "tags": [
          "breeder: buyer comms"
        ],
        "summary": "Send a direct message to one buyer.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "buyerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "subject": {
                    "type": "string",
                    "minLength": 1
                  },
                  "body": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "subject",
                  "body"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "created"
          },
          "404": {
            "description": "buyer not found"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/breeder/buyers/messages": {
      "get": {
        "tags": [
          "breeder: buyer comms"
        ],
        "summary": "Buyer message history (filter by buyerId / litterId).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "buyerId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "litterId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "messages": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    }
                  },
                  "required": [
                    "messages"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/breeder/buyers/update-pack/subscriptions": {
      "get": {
        "tags": [
          "breeder: buyer comms"
        ],
        "summary": "Weekly update-pack subscriptions.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "subscriptions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    }
                  },
                  "required": [
                    "subscriptions"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/breeder/buyers/update-pack/subscribe": {
      "post": {
        "tags": [
          "breeder: buyer comms"
        ],
        "summary": "Subscribe a buyer to a puppy’s weekly update pack.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "buyerId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "puppyId": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "buyerId",
                  "puppyId"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "created"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/breeder/buyers/update-pack/subscriptions/{id}": {
      "patch": {
        "tags": [
          "breeder: buyer comms"
        ],
        "summary": "Activate / deactivate an update-pack subscription.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "active": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "active"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ok"
          },
          "404": {
            "description": "not found"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/breeder/buyers/puppies/{pupId}/go-home-pack": {
      "get": {
        "tags": [
          "breeder: buyer comms"
        ],
        "summary": "Assembled go-home pack for a puppy.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "pupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          },
          "404": {
            "description": "puppy not found"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/breeder/buyers/update-pack/run": {
      "post": {
        "tags": [
          "breeder: buyer comms"
        ],
        "summary": "Run the weekly update-pack sweep now.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sent": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "sent"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    }
  }
}